refactor: rename semester-related classes and update package structure to improve organization
