+++ src/main/java/org/ject/support/domain/recruit/repository/SemesterRepository.java
@@ -1,7 +1,6 @@
-package org.ject.support.domain.semseter;
+package org.ject.support.domain.recruit.repository;
 
 import org.ject.support.domain.recruit.domain.Semester;
-import org.ject.support.domain.semseter.repository.SemesterQueryRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface SemesterRepository extends JpaRepository<Semester, Long>, SemesterQueryRepository {
